% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotClonalScores.R
\name{plotClonalScores}
\alias{plotClonalScores}
\title{Show the distribution of clonal scores}
\usage{
plotClonalScores(
  TCRanking,
  title = NULL,
  subtitle = NULL,
  palette = NULL,
  labelSize = 2,
  textSize = 7
)
}
\arguments{
\item{TCRanking}{Dataframe including clonal scores and ranking.
Could be obtained from \code{TCRanker()}}

\item{title}{title of the plot}

\item{palette}{Vector of hex color code, optional.}

\item{labelSize}{Numeric, font size of clonal scores}

\item{textSize}{Numeric, font size of clonotype sequence}
}
\value{
Plot the clonal ranking of all clonotypes against their size
}
\description{
Plot the clonal ranking of all clonotypes against their size,
including their clonal scores.
}
\examples{
plotClonalScores(TCRanking)
}
